+++ src/core/span_pipeline/finders/email.rs
@@ -6,7 +6,7 @@ use crate::core::config::EmailConfig;
 
 use super::super::span::{Collector, Finder};
 
-#[derive(Debug)]
+#[derive(Debug, Clone)]
 pub(crate) struct EmailFinder {
     regex: Regex,
     config: EmailConfig,