+++ src/core/span_pipeline/finders/json.rs
@@ -4,7 +4,7 @@ use crate::core::config::JsonConfig;
 
 use super::super::span::{Collector, Finder};
 
-#[derive(Debug)]
+#[derive(Debug, Clone)]
 pub(crate) struct JsonFinder {
     config: JsonConfig,
 }