+++ src/core/span_pipeline/finders/quote.rs
@@ -4,7 +4,7 @@ use crate::core::config::QuoteConfig;
 
 use super::super::span::{Collector, Finder};
 
-#[derive(Debug)]
+#[derive(Debug, Clone)]
 pub(crate) struct QuoteFinder {
     config: QuoteConfig,
 }